pyproject.toml
[tool.black]
line-length = 119
# include =
exclude = '''
/(
    \.git
  | __pycache__
  | .pytest_cache
  | .venv
  | build
  | dist
  | code/library/books/migrations/*
  | code/blogapi/posts/migrations/*
  | code/todo/backend/todos/migrations/*
  | ^.*\b(migrations)\b.*$
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
skip = ["*.gitignore"]
extend_skip = ["*.md", "*.json"]
line_length = 119

[tool.poetry]
name = "django_api"
version = "0.1.0"
description = "A Django REST Framework(DRF) sandbox"
authors = ["Kevin Bowen <kevin.bowen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.50.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
drf-yasg = "^1.20.0"
environs = {extras = ["django"], version = "^9.5.0"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
